requirements.txt
asgiref==3.5.2
Django==4.0.5
django-countries==7.3.2
django-filter==22.1
django-richtextfield==1.6.1
djangorestframework==3.13.1
Markdown==3.4.1
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
